SystemUI: show date on 2 lines in status bar
This allows showing of the full day of the week, and removes
the blank space above the date (which always felt strange to
me when clicking it to access the calendar).
It also fixes a spacing issue on smaller screen devices that
causes the quick settings button to be partially chopped off
when the AM/PM clock settings are enabled.

Change-Id: I715b88ac16b9180f0dc7ce3bf47a1f9198295cf9
Before: http://imgur.com/UzkFe
After: http://imgur.com/UXoBJ