WIP: Add active display to ChameleonOS [1/2]
This is the ChameleonOS implementation of the active
display feature, similar to that found on the Moto X.

When enabled the screen will turn on and display the
incoming notification.  The user can then open the app
for the notification, dismiss the notification or unlock
their device normally.

Additional notifications, if any exist, will be displayed
below the ring notification and can be viewed by touching
the notification icon.  The user can simply slide their
finger across these other notification to quickly view
them.  The last one touched will remain in the ring and
the user can then perform any of the actions listed above.

PS2-PS12:
All the changes that brought us to this current implementation.

Change-Id: I51c17cec33b65dc60111a10c5ca2c449596ebe63

Conflicts:

	core/java/android/provider/Settings.java
	core/res/res/values/dimens.xml
	core/res/res/values/symbols.xml
	packages/SystemUI/AndroidManifest.xml
	packages/SystemUI/res/values/dimens.xml
	packages/SystemUI/res/values/styles.xml
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/tablet/TabletStatusBar.java